+++ geograpix/src/App.js
@@ -18,6 +18,6 @@ function App() {
       </section>
     </div>
   );
-}
+};
 
 export default App;
