+++ geograpix/src/components/Map.js
@@ -1,14 +1,14 @@
 import React, { useState } from 'react';
 import ReactMapGL from 'react-map-gl';
-import {connect} from 'react-redux';
+import { connect } from 'react-redux';
 
 import PlotIcon from './PlotIcon';
 
 const Map = (props) => {
 
     const [viewport, setViewport] = useState({
-        latitude: 32.173114,
-        longitude: -95.773993,
+        latitude: 33.9243117,
+        longitude: -118.1327898,
         zoom: 10,
         width: '90vw',
         height: '90vh',
@@ -48,7 +48,7 @@ const Map = (props) => {
 
 const mapStateToProps = state => {
     return{
-        pictureInfo: state.pictureInfo
+        pictureInfo: state.maps.pictureInfo
     }
 }
