+++ geograpix/src/scss/index.scss
@@ -17,5 +17,5 @@ code {
   font-family: $font-code-stack;
 }
 
-@import './App';
+@import './registration';
 @import './marker';
