+++ geograpix/src/scss/registration.scss
@@ -6,7 +6,6 @@
   top: 0;
   left: 0;
   z-index: 10;
-  pointer-events: none;
 }
 
 .modal {
@@ -87,7 +86,7 @@
 
     button {
       width: 100%;
-      height: 40px;
+      height: 54px;
       margin-top: 30px;
       padding: 0 10px;
       position: relative;
