+++ geograpix/src/utilities/axiosWithAuth.js
@@ -1,0 +1,11 @@
+import axios from 'axios';
+const token = localStorage.getItem('token');
+
+export const axiosWithAuth = () => {
+  return axios.create({
+    headers: {
+      'Content-Type': 'application/json',
+      'Authorization': token
+    }
+  });
+};
